Refactor to use fxml
