+++ src/main/java/DialogBox.java
@@ -1,5 +1,9 @@
+import java.io.IOException;
+
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
+import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
 import javafx.geometry.Pos;
 import javafx.scene.Node;
 import javafx.scene.control.Label;
@@ -8,13 +12,24 @@ import javafx.scene.image.ImageView;
 import javafx.scene.layout.HBox;
 
 public class DialogBox extends HBox {
+    @FXML
     private Label label;
+    @FXML
     private ImageView imageView;
 
     private DialogBox(String text, Image picture) {
-        this.label = new Label(text);
-        this.imageView = new ImageView(picture);
-        this.getChildren().addAll(label, imageView);
+        try {
+            FXMLLoader fxmlLoader = new FXMLLoader(MainWindow.class.getResource("/view/DialogBox.fxml"));
+            fxmlLoader.setController(this);
+            fxmlLoader.setRoot(this);
+            fxmlLoader.load();
+        } catch (IOException e) {
+            System.out.println(e.getMessage());
+        }
+
+        this.label.setText(text);
+        this.imageView.setImage(picture);
+
     }
 
     public static DialogBox getHumanDialog(String text, Image picture) {
@@ -30,16 +45,10 @@ public class DialogBox extends HBox {
     }
 
     private void styleAsUser() {
-        this.label.setWrapText(true);
-        this.imageView.setFitWidth(100.0);
-        this.imageView.setFitHeight(100.0);
         this.setAlignment(Pos.TOP_RIGHT);
     }
 
     private void styleAsBot() {
-        this.label.setWrapText(true);
-        this.imageView.setFitWidth(100.0);
-        this.imageView.setFitHeight(100.0);
         this.setAlignment(Pos.TOP_LEFT);
 
         ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
